cambios menores a componentes
