+++ Client/src/components/Navbar/Navbar.css
@@ -7,7 +7,7 @@
     color: black;
     padding: 10px;
     text-decoration: none;
-    font-size: 20px;
+    font-size: 15px;
   }
 
   .buttons:hover {
