+++ Client/src/main.jsx
@@ -3,17 +3,14 @@ import ReactDOM from 'react-dom/client'
 import App from './App.jsx'
 import './index.css'
 import { BrowserRouter } from "react-router-dom"
-import { SearchProvider } from './services/SearchContext.jsx'
 import { ProductProvider } from './services/ProductContext.jsx'
 
 ReactDOM.createRoot(document.getElementById('root')).render(
   <BrowserRouter>
-    <SearchProvider>
-    <ProductProvider>  
+    <ProductProvider>
       <React.StrictMode>
         <App />
       </React.StrictMode>
-      </ProductProvider >
-    </SearchProvider>
+    </ProductProvider >
   </BrowserRouter>
 )
